requirements.txt
django

djangorestframework
python-dotenv
psycopg2-binary
django-phonenumber-field
pillow
ipython
django-filter
djangorestframework-simplejwt

# Для стилизации импортов
isort
